# src/main/resources/static/openapi.yaml
openapi: 3.0.3
info:
  title: Erudit API
  version: 1.0.0
  description: Контракт всех реализованных и запланированных REST-эндпоинтов. Операции с пометкой «Запланировано»
    появятся после выполнения соответствующих задач; Try it out для них пока вернёт 404.
servers:
- url: /
...
paths:
...
  /api/v1/content:
    get:
      tags:
      - Content
      operationId: listContent
      summary: Каталог контента
      description: Реализовано в задаче 18. Возвращает только опубликованный контент; размер страницы не больше 50.
      x-task-id: 18
      x-implementation-status: implemented
      security: []
      parameters:
      - *id002
      - *id003
      - name: categoryId
        in: query
        required: false
        schema: *id001
      - name: type
        in: query
        required: false
        schema:
          type: string
      - name: difficulty
        in: query
        required: false
        schema:
          type: string
      - name: premium
        in: query
        required: false
        schema: &id008
          type: boolean
      - name: query
        in: query
        required: false
        schema:
          type: string
      - name: sort
        in: query
        required: false
        schema:
          type: string
      responses:
        '200':
          description: Успешный ответ
          headers:
            X-Total-Count:
              description: Общее количество элементов с учётом фильтров
              schema:
                type: integer
                format: int64
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ContentListResponse'
        default:
          description: Ошибка
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    post:
      tags:
      - Content
      operationId: createContent
      summary: Создать контент
      description: Реализовано в задаче 15. Требуется роль ADMIN.
      x-task-id: 15
      x-implementation-status: implemented
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ContentUpsertRequest'
      responses:
        '201':
          description: Успешный ответ
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ContentItemResponse'
        default:
          description: Ошибка
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
      x-required-role: ADMIN
  /api/v1/content/categories:
    get:
      tags:
      - Content
      operationId: listCategories
      summary: Категории
      description: Реализовано в задаче 18. Размер страницы не больше 50.
      x-task-id: 18
      x-implementation-status: implemented
      security: []
      parameters: *id004
      responses:
        '200':
          description: Успешный ответ
          headers:
            X-Total-Count:
              description: Общее количество категорий
              schema:
                type: integer
                format: int64
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CategoryListResponse'
        default:
          description: Ошибка
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/v1/content/formats:
    get:
      tags:
      - Content
      operationId: listFormats
      summary: Форматы контента
      description: Реализовано в задаче 18. Справочник ограничен enum ContentType.
      x-task-id: 18
      x-implementation-status: implemented
      security: []
      responses:
        '200':
          description: Успешный ответ
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FormatList'
        default:
          description: Ошибка
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/v1/content/feed:
    get:
      tags:
      - Content
      operationId: getContentFeed
      summary: Персональная лента
      description: Запланировано в задаче 19. До реализации этот маршрут вернёт 404.
      x-task-id: 19
      x-implementation-status: planned
      parameters: *id004
      responses:
        '200':
          description: Успешный ответ
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ContentListResponse'
        default:
          description: Ошибка
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/v1/content/{id}:
    get:
      tags:
      - Content
      operationId: getContent
      summary: Получить контент
      description: Реализовано в задаче 15. Неопубликованный контент доступен только ADMIN.
      x-task-id: 15
      x-implementation-status: implemented
      security: []
      parameters:
      - name: id
        in: path
        required: true
        description: Идентификатор ресурса
        schema: *id001
      responses:
        '200':
          description: Успешный ответ
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ContentItemResponse'
        default:
          description: Ошибка
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    put:
      tags:
      - Content
      operationId: updateContent
      summary: Обновить контент
      description: Реализовано в задаче 15. Требуется роль ADMIN.
      x-task-id: 15
      x-implementation-status: implemented
      parameters:
      - name: id
        in: path
        required: true
        description: Идентификатор ресурса
        schema: *id001
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ContentUpsertRequest'
      responses:
        '200':
          description: Успешный ответ
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ContentItemResponse'
        default:
          description: Ошибка
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
      x-required-role: ADMIN
    delete:
      tags:
      - Content
      operationId: archiveContent
      summary: Архивировать контент
      description: Реализовано в задаче 15. Выполняет мягкое удаление со статусом ARCHIVED. Требуется роль ADMIN.
      x-task-id: 15
      x-implementation-status: implemented
      parameters:
      - name: id
        in: path
        required: true
        description: Идентификатор ресурса
        schema: *id001
      responses:
        '204':
          description: Операция выполнена
        default:
          description: Ошибка
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
      x-required-role: ADMIN
  /api/v1/content/{id}/quiz-cards:
    get:
      tags:
      - Content
      operationId: getContentQuizCards
      summary: Карточки вопросов по контенту
      description: Реализовано в задаче 17. Для неопубликованного контента требуется роль ADMIN.
      x-task-id: 17
      x-implementation-status: implemented
      parameters:
      - name: id
        in: path
        required: true
        description: Идентификатор ресурса
        schema: *id001
      responses:
        '200':
          description: Успешный ответ
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/QuizCardListResponse'
        default:
          description: Ошибка
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/v1/content/{id}/progress:
    get:
      tags:
      - Content
      operationId: getContentProgress
      summary: Прогресс по контенту
      description: Реализовано в задаче 22. Возвращает прогресс текущего пользователя.
      x-task-id: 22
      x-implementation-status: implemented
      parameters:
      - name: id
        in: path
        required: true
        description: Идентификатор ресурса
        schema: *id001
      responses:
        '200':
          description: Успешный ответ
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ContentProgressResponse'
        default:
          description: Ошибка
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/v1/content/{id}/view:
    post:
      tags:
      - Content
      operationId: markContentViewed
      summary: Отметить просмотр
      description: Реализовано в задаче 22. Идемпотентно отмечает первый просмотр.
      x-task-id: 22
      x-implementation-status: implemented
      parameters:
      - name: id
        in: path
        required: true
        description: Идентификатор ресурса
        schema: *id001
      responses:
        '200':
          description: Успешный ответ
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ContentProgressResponse'
        default:
          description: Ошибка
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/v1/content/{id}/complete:
    post:
      tags:
      - Content
      operationId: completeContent
      summary: Завершить изучение
      description: Реализовано в задаче 22. Идемпотентно завершает изучение контента.
      x-task-id: 22
      x-implementation-status: implemented
      parameters:
      - name: id
        in: path
        required: true
        description: Идентификатор ресурса
        schema: *id001
      responses:
        '200':
          description: Успешный ответ
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ContentProgressResponse'
        default:
          description: Ошибка
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    ApiError:
      type: object
      properties:
        code:
          type: string
        message:
          type: string
      required:
      - code
      - message
    ErrorResponse:
      type: object
      properties:
        error:
          $ref: '#/components/schemas/ApiError'
      required:
      - error
    PageMetadata:
      type: object
      properties:
        page: &id007
          type: integer
          format: int32
        size: *id007
        totalElements:
          type: integer
          format: int64
        totalPages: *id007
      required:
      - page
      - size
      - totalElements
      - totalPages
...
    ContentItem:
      type: object
      properties:
        id: *id001
        categoryId: *id001
        type:
          type: string
          enum: [ARTICLE, FACT_CARD, AUDIO, VIDEO, COMBINED, QUIZ_BLOCK]
        title:
          type: string
          minLength: 1
          maxLength: 250
        description:
          type: string
          maxLength: 2000
        body:
          type: string
        mediaUrl:
          type: string
        difficulty:
          type: string
        estimatedMinutes: *id007
        authorId:
          type: string
        status:
          type: string
        tags:
          type: array
          items:
            type: string
        premiumLocked: *id008
      required:
      - id
      - type
      - title
      - status
    Category:
      type: object
      properties:
        id: *id001
        name:
          type: string
      required:
      - id
      - name
    FormatList:
      type: object
      properties:
        data:
          type: array
          items:
            type: string
      required:
      - data
...
    QuizCard:
      type: object
      properties:
        id: *id001
        fact:
          type: string
          description: Факт из исходного материала, на котором основан вопрос
        question:
          type: string
        answers:
          type: array
          items:
            type: string
      required:
      - id
      - fact
      - question
      - answers
    ContentProgress:
      type: object
      properties:
        contentId: *id001
        status:
          type: string
        viewedAt: *id005
        completedAt: *id005
      required:
      - contentId
      - status
...
    ContentUpsertRequest:
      type: object
      properties:
        categoryId: *id001
        type:
          type: string
          enum: [ARTICLE, FACT_CARD, AUDIO, VIDEO, COMBINED, QUIZ_BLOCK]
        title:
          type: string
        description:
          type: string
        body:
          type: string
        mediaUrl:
          type: string
          maxLength: 1000
        difficulty:
          type: string
          enum: [BEGINNER, INTERMEDIATE, ADVANCED]
        estimatedMinutes:
          type: integer
          format: int32
          minimum: 1
        premiumLocked:
          type: boolean
          default: false
        tags:
          type: array
          uniqueItems: true
          items:
            type: string
            minLength: 1
            maxLength: 100
      required:
      - categoryId
      - type
      - title
...
    ContentItemResponse:
      type: object
      properties:
        data:
          $ref: '#/components/schemas/ContentItem'
      required:
      - data
    ContentListResponse:
      type: object
      properties:
        data:
          type: array
          items:
            $ref: '#/components/schemas/ContentItem'
        pagination:
          $ref: '#/components/schemas/PageMetadata'
      required:
      - data
    CategoryListResponse:
      type: object
      properties:
        data:
          type: array
          items:
            $ref: '#/components/schemas/Category'
        pagination:
          $ref: '#/components/schemas/PageMetadata'
      required:
      - data
    QuizCardListResponse:
      type: object
      properties:
        data:
          type: array
          items:
            $ref: '#/components/schemas/QuizCard'
        pagination:
          $ref: '#/components/schemas/PageMetadata'
      required:
      - data
    ContentProgressResponse:
      type: object
      properties:
        data:
          $ref: '#/components/schemas/ContentProgress'
      required:
      - data
...
security:
- bearerAuth: []